+++ com.unity.formats.alembic/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 ### Fixed
+- Fixed a bug that caused an error when manually adding an AlembicStreamPlayer Component.
 
 ## [2.2.0] - 2021-06-10
 ### Added